getters: Simplify get_commodity_map() and rename get_commodity_directives()
All callers of get_commodity_map() only care for the metadata
associated to commodities via explicit 'commodity' directives. Thus
there is no need to try to collect commodities used elsewhere in the
ledger. To clarify the purpose of the function and to avoid possible
sublte breakage in external code, rename the function to
get_commodity_directives().