Remove bin/bean-query
The executable is created by setuptools on install. Alternatively the
tool can be run from the source directory with "python -m beanquery".

Martin is probably the only one that does not install this with pip or
equivalent. For him I have a tweak to the etc/env file to add a
bean-query alias that invokes "python -m beanquery".