Make tests self contained
Embed the test ledger into the test harnesses. Make the test ledger
much smaller and thus the tests easier to verify. This also allows to
get rid of the example.beancount file and of the code required to find
it and makes the tests work when the package is installed.