Add BQL type casting functions
The str() function used to return the repr of its argument. Rename it
to repr(). The date() function on str argument used to parse the date
liberally. Rename it parse_date() and use strict parsing in the date()
type casting function. Add another form of parse_date() that accepts a
format as first argument.