render: Fix inventory renderer
The assumptions in the former implementation were too simplistic and
caused positions with the same units commodity but different costs to
be lost. Give up on the nice tabular format when there are too many
distinct positions to be rendered. In this case the estimated column
width becomes really a worst case estimate and unnecessary trailing
whitespace may be present. The only way to fix this properly would be
a three pass rendering algorithm. However fixing a cosmetic issue in a
table that is going anyhow to do not look neat is not worth the cost.