Remove partial support for the FLATTEN query option
The implementation was broken and incomplete. It is not clear what to
do when two columns in the results set contain two inventory columns.
The FLATTEN query option also partially overlaps with the "expand"
option in the table renderer. It has been broken for quite some time
and there have not been many complaints. Just nuke it. If someone
comes up with a sensible design it will be easy to reinstate it.