parser: Use dedicated AST node for the != operator
Add a NotEqual() AST node to represent the != comparison operator
replacing the compound Not(Equal()) form used till now.

This improves performance slightly as it reduces the number of
function calls required to execute the query. Most importantly, it
regularizes the implementation of the binary operations. This will be
useful when switching to a different parser generator.