Implement support for SELECT ... FROM #table expressions
The existing BQL syntax allows for arbitrary expressions on the FROM
clause thus we need a way to disambiguate between a table name and a
field reference. Restrict table names to valid identifiers prefixed
with the # character.

The only available tables are the 'entries' and 'postings' tables and
the null table (obtained with 'SELECT ... FROM #') which is useful to
evaluate simple BQL expressions.