Restructure internals and API
``beanquery.connect()`` becomes the main entry point. It takes an URI
pointing to a data source and returns a Connection() class with an
``execute()`` method that takes a query string and returns the query
results.  The only data source scheme defined so far is ``beancount:``
which is expected to be followed by the path to a Beancount ledger.

This removes the use of a module global variable to keep the list of
defined tables and continues the effort toward the generalization of
beanquery to work with data sources other than beancount and toward an
API somehow similar to the Python's DB-API 2.0.

The main entry point of the old API (beanquery.query.run_query() aka
beancount.query.run_query() before the project split) is preserved for
backward compatibility.  As no other interface was clearly meant as
public API, no particular effort has been put in preserving backward
compatibility for other functions or classes.