Adjust filenamee and lineno columns definition in the postings table
These columns are documented to return the "directive location", which
should be interpreted as the location of the posting definition (the
location column indeed returns filename:lineno for the posting).
However, the implementation returned the transaction location.

With the recent addition of the meta and entry columns, the location
of the transaction definition can be obtained with entry.meta['filename']
and entry.meta['lineno'].

Adjust the filename and lineno columns in the postings table to refer
to the posting location and not to the transaction location.

Fixes #96.