Use dateutil.parser directly
Remove indirection via beancount.utils.date_utils. This makes the
dependency on python-dateutil explicit and will allow to remove the
utility code from Beancount.

The one argument form of the parse_date() BQL function that tries to
guess the date format does not seem a great idea as the format used
for the conversion depends on the input string, thus it could
interpret different rows differently. It should be deprecated and
eventually removed.