env: Rewrite most BQL functions taking a reference to the data row
Make them take a reference to the evaluation context instead, and look
up the required information from other tables. This allows to simplify
the structure of the row returned by the postings table.

meta(), entry_meta(), any_meta(), and has_account() cannot be updated
to the new schema as they access the entry relative to the current
posting. These functions will be deprecated and replaced by other
constructs.