Implement has_account() as AST substitution
The AST is rewritten to contain an explicit reference to the accounts
column. This removes the need to pass the current row object (which is
of the expected kind only when iterating the postings or entries
tables) to the function evaluation. This removes the last function
implementation that uses the current row.