Implement a "?~" operator
It works like the regular expression match operator "~" but with the
arguments in the reversed order (to make it more useful in combination
with the ALL and ANY operators). Furthermore, it does not default to
case insensitive matching. The regular expression inline flag (?i) can
be used to perform case insensitive matching.

This does not follow the SQL standard nor there is precedent for it in
any SQL-like query language that I know of. The operator name is not
set in stone and can be adapted to any prior art.