pyproject.toml: use SPDX identifier to describe license
Rationale: by specifying the license as a file (like before), the full text of the license is included in the package metadata, and shown in various places, e.g., in the output of `pip show`, which is quite cumbersome. With this change, we use the alternative way of specifying license (namely: using a standard SPDX identifier), which makes license metadata concise and, incidentally, also machine readable.