Render ProgrammingError, CompilationError and ParserError alway user-friendly
Change `render_exception()` so that it does not show a stack trace if no
AST node information is provided to CompilationError.

Also, do not print a stack trace for `ProgrammingError`s, as these are
also errors committed by the user and therefore should have a nice error
message without a stack trace.